[ISSUE-1254] Move firebase APIs to jmap-extensions-api
