+++ tmail-backend/jmap/extensions-api/src/main/java/com/linagora/tmail/james/jmap/firebase/FirebaseSubscriptionRepository.java
