+++ tmail-backend/jmap/extensions/src/main/scala/com/linagora/tmail/james/jmap/model/FirebaseSubscription.scala
@@ -1,7 +1,6 @@
 package com.linagora.tmail.james.jmap.model
 
 import java.time.ZonedDateTime
-import java.util.UUID
 
 import cats.data.Validated
 import cats.instances.list._
@@ -11,7 +10,6 @@ import com.linagora.tmail.james.jmap.model.FirebaseSubscriptionPatchObject.updat
 import com.linagora.tmail.james.jmap.model.FirebaseSubscriptionUpdateFailure.LOGGER
 import eu.timepit.refined.auto._
 import org.apache.james.jmap.api.change.TypeStateFactory
-import org.apache.james.jmap.api.model.ExpireTimeInvalidException.TIME_FORMATTER
 import org.apache.james.jmap.api.model.TypeName
 import org.apache.james.jmap.core.Id.Id
 import org.apache.james.jmap.core.Properties.toProperties
@@ -22,104 +20,6 @@ import org.slf4j.{Logger, LoggerFactory}
 import play.api.libs.json.{JsArray, JsObject, JsPath, JsString, JsValue, JsonValidationError}
 
 import scala.util.{Failure, Success, Try}
-
-object FirebaseSubscriptionId {
-  def generate(): FirebaseSubscriptionId = FirebaseSubscriptionId(UUID.randomUUID)
-
-  def liftOrThrow(unparsedId: UnparsedFirebaseSubscriptionId): Either[IllegalArgumentException, FirebaseSubscriptionId] =
-    liftOrThrow(unparsedId.id.value)
-
-  def liftOrThrow(value: String): Either[IllegalArgumentException, FirebaseSubscriptionId] =
-    Try(UUID.fromString(value))
-      .map(value1 => FirebaseSubscriptionId(value1))
-      .toEither
-      .left.map(e => new IllegalArgumentException("FirebaseSubscriptionId is invalid", e))
-}
-
-case class FirebaseSubscriptionId(value: UUID) {
-  def serialize: String = value.toString
-
-  def asUnparsedFirebaseSubscriptionId: UnparsedFirebaseSubscriptionId =
-    UnparsedFirebaseSubscriptionId(Id.validate(serialize).toOption.get)
-}
-
-case class DeviceClientId(value: String) extends AnyVal
-
-case class FirebaseToken(value: String) extends AnyVal
-
-case class FirebaseSubscriptionExpiredTime(value: ZonedDateTime) {
-  def isAfter(date: ZonedDateTime): Boolean = value.isAfter(date)
-
-  def isBefore(date: ZonedDateTime): Boolean = value.isBefore(date)
-}
-
-case class FirebaseSubscriptionCreationRequest(deviceClientId: DeviceClientId,
-                                               token: FirebaseToken,
-                                               expires: Option[FirebaseSubscriptionExpiredTime] = None,
-                                               types: Seq[TypeName]) {
-
-  def validate: Either[IllegalArgumentException, FirebaseSubscriptionCreationRequest] =
-    validateTypes
-
-  private def validateTypes: Either[IllegalArgumentException, FirebaseSubscriptionCreationRequest] =
-    if (types.isEmpty) {
-      scala.Left(new IllegalArgumentException("types must not be empty"))
-    } else {
-      Right(this)
-    }
-}
-
-object FirebaseSubscription {
-  val EXPIRES_TIME_MAX_DAY: Int = 7
-  val allProperties: Properties = Properties("id", "deviceClientId", "expires", "types")
-  val idProperty: Properties = Properties("id")
-
-  def from(creationRequest: FirebaseSubscriptionCreationRequest,
-           expireTime: FirebaseSubscriptionExpiredTime): FirebaseSubscription =
-    FirebaseSubscription(id = FirebaseSubscriptionId.generate(),
-      deviceClientId = creationRequest.deviceClientId,
-      token = creationRequest.token,
-      expires = expireTime,
-      types = creationRequest.types)
-}
-
-case class FirebaseSubscription(id: FirebaseSubscriptionId,
-                                deviceClientId: DeviceClientId,
-                                token: FirebaseToken,
-                                expires: FirebaseSubscriptionExpiredTime,
-                                types: Seq[TypeName]) {
-  def withTypes(types: Seq[TypeName]): FirebaseSubscription = copy(types = types)
-
-  def withExpires(expires: FirebaseSubscriptionExpiredTime): FirebaseSubscription = copy(expires = expires)
-}
-
-case class FirebaseSubscriptionNotFoundException(id: FirebaseSubscriptionId) extends RuntimeException
-
-case class ExpireTimeInvalidException(expires: ZonedDateTime, message: String) extends IllegalStateException(s"`${expires.format(TIME_FORMATTER)}` $message")
-
-case class DeviceClientIdInvalidException(deviceClientId: DeviceClientId, message: String) extends IllegalArgumentException(s"`${deviceClientId.value}` $message")
-
-case class TokenInvalidException(message: String) extends IllegalArgumentException(message)
-
-case class MissingOrInvalidFirebaseCredentialException(message: String) extends IllegalArgumentException(message)
-
-case class UnparsedFirebaseSubscriptionId(id: Id)
-
-case class FirebaseSubscriptionIds(list: List[UnparsedFirebaseSubscriptionId])
-
-case class FirebaseSubscriptionGetRequest(ids: Option[FirebaseSubscriptionIds],
-                                          properties: Option[Properties]) extends WithoutAccountId {
-
-  def validateProperties: Either[IllegalArgumentException, Properties] =
-    properties match {
-      case None => Right(FirebaseSubscription.allProperties)
-      case Some(value) =>
-        value -- FirebaseSubscription.allProperties match {
-          case invalidProperties if invalidProperties.isEmpty() => Right(value ++ FirebaseSubscription.idProperty)
-          case invalidProperties: Properties => Left(new IllegalArgumentException(s"The following properties [${invalidProperties.format()}] do not exist."))
-        }
-    }
-}
 
 object FirebaseSubscriptionGetResponse {
   def from(list: Seq[FirebaseSubscription], requestIds: Option[FirebaseSubscriptionIds]): FirebaseSubscriptionGetResponse =
@@ -159,9 +59,9 @@ object FirebaseSubscriptionCreationParseException {
 case class FirebaseSubscriptionCreationParseException(setError: SetError) extends Exception
 
 object FirebaseSubscriptionCreation {
-  private val serverSetProperty: Set[String] = Set("id")
-  private val assignableProperties: Set[String] = Set("deviceClientId", "token", "expires", "types")
-  private val knownProperties: Set[String] = assignableProperties ++ serverSetProperty
+  val serverSetProperty: Set[String] = Set("id")
+  val assignableProperties: Set[String] = Set("deviceClientId", "token", "expires", "types")
+  val knownProperties: Set[String] = assignableProperties ++ serverSetProperty
 
   def validateProperties(jsObject: JsObject): Either[FirebaseSubscriptionCreationParseException, JsObject] =
     (jsObject.keys.intersect(serverSetProperty), jsObject.keys.diff(knownProperties)) match {
